+++ Block/System/Config/ColorPicker.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * @author @haihv433
+ * @copyright Copyright (c) 2020 Goomento (https://store.goomento.com)
+ * @package Goomento_Base
+ * @link https://github.com/Goomento/Base
+ */
+
+namespace Goomento\Base\Block\System\Config;
+
+/**
+ * Class ColorPicker
+ * @package Goomento\Base\Block\System\Config
+ */
+class ColorPicker extends \Magento\Config\Block\System\Config\Form\Field
+{
+    /**
+     * @param \Magento\Backend\Block\Template\Context $context
+     * @param array $data
+     */
+    public function __construct(
+        \Magento\Backend\Block\Template\Context $context,
+        array $data = []
+    ) {
+        parent::__construct($context, $data);
+    }
+
+    /**
+     * @param \Magento\Framework\Data\Form\Element\AbstractElement $element
+     * @return string
+     */
+    protected function _getElementHtml(\Magento\Framework\Data\Form\Element\AbstractElement $element)
+    {
+        $html = $element->getElementHtml();
+        $value = $element->getData('value');
+
+        $html .= '<script type="text/javascript">
+        require(["jquery","jquery/colorpicker/js/colorpicker"], function ($) {
+            $(document).ready(function () {
+                let $el = $("#' . $element->getHtmlId() . '");
+                $el.css("backgroundColor", "' . $value . '");
+                $el.ColorPicker({
+                    color: "' . $value . '",
+                    onChange: function (hsb, hex, rgb) {
+                        $el.css("backgroundColor", "#" + hex).val("#" + hex);
+                    }
+                });
+            });
+        });
+        </script>';
+        return $html;
+    }
+}