+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Magento 2 Goomento Base
 
+## 1.0.2: Enhance System Configuration 
+### Features:
+
+Added:
+- Color Picker for System Configuration
+- Frontend config provider
+
 ## 1.0.1: Dynamic Instance Manager calling added
 ### Features: